Update Product, Delete Product Image, ProductPolicy

+++ app/Http/Controllers/Store/ProductController.php
@@ -4,13 +4,18 @@ namespace App\Http\Controllers\Store;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Store\ProductCreateRequest;
+use App\Http\Requests\Store\ProductUpdateRequest;
+use App\Models\Image;
 use App\Models\ProductCondition;
 use App\Repositories\UserRepository;
 use App\Services\CategoryService;
 use App\Services\ProductService;
 use Illuminate\Contracts\View\View;
+use Illuminate\Database\QueryException;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
 
 class ProductController extends Controller
 {
@@ -45,7 +50,7 @@ class ProductController extends Controller
     {      
         $data=$request->validated();
         $this->productService->createProduct($data);
-        return redirect()->route('products.index')->with('success','Product deleted!');;
+        return redirect()->route('products.index')->with('success','Product deleted!');
     }
 
     /**
@@ -61,15 +66,24 @@ class ProductController extends Controller
      */
     public function edit(string $id)
     {
-        //
+        $this->authorize('update',$this->productService->getProduct($id));
+        return view('store.product.edit',[
+            'user'=>$this->user->getUser(),
+            'product'=>$this->productService->getProduct($id),
+            'categories'=>$this->categoryService->getAllCategory(),
+            'conditions'=>ProductCondition::all()
+        ]);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(ProductUpdateRequest $request,$id)
     {
-        //
+        $this->authorize('update',$this->productService->getProduct($id));
+        $data=$request->validated();
+        $this->productService->updateProduct($id,$data);
+        return redirect()->route('products.index')->with('success','Product updated!');
     }
 
     /**
@@ -77,7 +91,22 @@ class ProductController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete',$this->productService->getProduct($id));
         $this->productService->deleteProduct($id);
         return redirect()->route('products.index')->with('success','Product deleted!');
+    }
+    
+    public function destroyImage($product,Image $image){
+        $this->authorize('delete',$this->productService->getProduct($product));
+        try{
+            DB::beginTransaction();
+            Storage::delete($image->url);
+            $image->products()->detach($product);
+            $image->delete();
+            DB::commit();
+        }catch(QueryException $e){
+            DB::rollBack();
+        }
+        return redirect()->route('products.edit',$product);
     }
 }
